% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acwri-package.R
\docType{package}
\name{acwri-package}
\alias{acwri}
\alias{acwri-package}
\title{acwri: Academic Writing Tools for R User}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of tools for academic writing in R. This package integrates various utilities to streamline the academic writing workflow, including document setup, formatting, and other helpful functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jixing475/acwri}
  \item Report bugs at \url{https://github.com/jixing475/acwri/issues}
}

}
\author{
\strong{Maintainer}: Jixing Liu \email{jixing475@gmail.com}

}
\keyword{internal}
